move start calculation to handler
